Remove file cache entry on IOException
The asynchronous removal of a CASFileCache entry's file backing must be
written back to the storage as a removal when it is observed. This
corrects a truth inconsistency between contains, used by
findMissingBlobs and get, used by the instance's getBlob. A stat check
with a ttl for a file is also available, but does not perfectly account
for all possible IOExceptions for a file.