Switch match interface to channeled listener
A match will receive several notifications over its lifetime:
Any number of pairs of onWaitStart, onWaitEnd, with multiple onWaitEnd
calls being possible. These indicate the start and end of a blocking
wait for notification of work.
An optional onOperationName event, if the work protocol supports it,
distinct from the availability of a full Operation. If the call returns
false, no further processing occurs and the match returns.
An onOperation event. The return should be an indicator as to its
successful processing.

Provide a layered requeuing implementation in the worker
implementations, and a poller for the match stage, which has been
observed to be a dispatched expiration trigger.