Retain QueuedOperations in CAS
In the interest of removing the requirement that the backplane retain
composite operations ready for execution, this change moves their
availability into the CAS, and provides a digest in both the queue entry
and the QueuedOperationMetadata for retrieving those results. This led
to a refactor of the prequeue/queue process to include structured
execute and queue entries separate from metadata.

The global operation failsafe routine was removed from the redis
backplane implementation to be replaced by an age-based behavior in a
forthcoming change. This simplified the stage-dependent reconnect
behavior for subscriptions, and will defer to an injected control system
for appropriate reinsertion and reaction of operations in failsafe.