MatchListener onEntry must accept null queueEntry
A backplane is fully capable of producing null queueEntry responses from
dispatchOperation, and the null operation response is an indicator that
a problem occurred and that the match must be retried. This response is
transferred to delegate listeners for RetryingMatchListeners, and
constitutes a matched response with a return from the match call.