Stream result content into exec storage for shard
Rather than load all of these results into memory in order to emplace
into the CAS, compute the digests and stream them into it. This required
the addition of the OutputStreamFactory to the ContentAddressableStorage
interface, which was on the eventual migration of all CAS
implementations to streaming.

Leaving the grpc CAS newOutput unimplemented for now.