Lock expiring key to prevent access of unlinked
When an entry is being expired, we must acquire its key lock in order to
prevent accesses from being recorded on it. This was particularly likely
when determining whether a key was still maintained after expiration,
but could occur as a race any time an entry access could occur, even
with proper locks. The digest expiration check is now rendered more
intelligent and harmless with respect to access.